+++ day8.py
@@ -8,68 +8,13 @@ data = removeNewLine(getData(__file__, 'test'))
 
 w = len(data[0])
 h = len(data)
-
 print(w, ' X ', h)
-
-#printLst(data)
 
 def fromStrToIntArray(str):
     out = []
     for c in str:
         out.append(int(c))
     return out
-
-def fromStrToIntNpArray(str):
-    out = []
-    for c in str:
-        out.append(int(c))
-    return np.asarray(out)
-
-def staircaseArray(arr):
-    indexArr = np.zeros(9)
-
-    for i in range(9):
-        indexArr[i] = arr.find(str(i))
-    return indexArr
-
-def getDifMatrix(data):
-    out = []
-    previousRow = np.zeros(w)
-
-    for line in data:
-        row = fromStrToIntNpArray(line) 
-        print(previousRow)
-        print(row)
-        dif = row > previousRow
-        print(dif)     
-        out.append(dif)
-        previousRow = row
-        #print(out)     
-
-    return out
-
-def getSingleVisibleTreeIndex(treeLine):
-    treeLine = fromStrToIntArray(treeLine)
-    maxTreeHeight = max(treeLine)
-    indexesOfMaxTrees = [i for (y, i) in zip(treeLine, range(len(treeLine))) if maxTreeHeight == y]
-
-    if len(indexesOfMaxTrees) > 1:
-        return -1
-    return indexesOfMaxTrees[0]
-
-def getVisibleTree(treeLine):
-    treeIndex = getSingleVisibleTreeIndex(treeLine)
-    treeLineBool = np.zeros(len(treeLine))
-    if treeIndex < 0:
-        return treeLineBool
-    treeLineBool[treeIndex] = 1
-    return treeLineBool
-
-def get2dArrayWithVisibleTrees(forrest):
-    array2dWithMaxTrees = []
-    for treeLine in forrest:
-        array2dWithMaxTrees.append(getVisibleTree(treeLine))
-    return np.asarray(array2dWithMaxTrees)
 
 def fromStrToNpData(data):
     npData = []
@@ -78,18 +23,10 @@ def fromStrToNpData(data):
     return np.asarray(npData)
 
 npData = fromStrToNpData(data)
-#printLst(data)
-rotData = np.rot90(npData)
 
-def getNumberOfVisibleTrees(forrest):
-    horizontalVisible = get2dArrayWithVisibleTrees(data)
-    verticalVisible = np.rot90(get2dArrayWithVisibleTrees(rotData), k=1, axes=(1,0))
-    return (np.logical_and(horizontalVisible, verticalVisible)).sum()
-
-def getVisibleTrees(treeLine):
-
+def getVisibleTreesInLine(treeLine):
     visibleTrees = []
-    visibleHeight = 0
+    visibleHeight = -1
     for t in treeLine:
         if t > visibleHeight:
             visibleHeight = t
@@ -98,13 +35,25 @@ def getVisibleTrees(treeLine):
             visibleTrees.append(0)
     return visibleTrees
 
-#print(getNumberOfVisibleTrees(data))
-npData = fromStrToNpData(data)
+def getVisibleTreesFromOneDirection(forrest):
+    boolForrest = []
+    for l in forrest: 
+        boolForrest.append(getVisibleTreesInLine(l))
+    return boolForrest
 
-for l in npData: 
-    print(l)
-    print(getVisibleTrees(l))
+def getAllVisibleTreesInForrest(forrest):
+    boolForrest4Directions = []
+    for direction in range(4):
+        boolForrest = getVisibleTreesFromOneDirection(np.rot90(forrest, direction+1))
+        boolForrest4Directions.append(np.rot90(boolForrest, -1*direction+1))
+    return np.asarray(boolForrest4Directions)
 
+
+x = getAllVisibleTreesInForrest(npData)
+bool0and1 = np.logical_and(x[0],x[1])
+bool2and3 = np.logical_and(x[2],x[3])
+boolAll = np.logical_and(bool0and1, bool2and3)
+print(boolAll.sum())
 
 
 # PART 2 ###################################################
